scripts/requirements.txt
python-dotenv
requests
psycopg2-binary
neo4j
